refactor: 다음 바이옴 네이밍 변경
