+++ android/app/src/main/java/poke/rogue/helper/presentation/biome/detail/nextbiomes/BiomeNextBiomesAdapter.kt
@@ -1,20 +1,20 @@
-package poke.rogue.helper.presentation.biome.detail.map
+package poke.rogue.helper.presentation.biome.detail.nextbiomes
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
-import poke.rogue.helper.databinding.ItemBiomeMapBinding
+import poke.rogue.helper.databinding.ItemBiomeNextBiomesBinding
 import poke.rogue.helper.presentation.biome.model.NextBiomeUiModel
 import poke.rogue.helper.presentation.util.view.ItemDiffCallback
 
-class BiomeMapAdapter :
-    ListAdapter<NextBiomeUiModel, BiomeMapViewHolder>(biomeComparator) {
+class BiomeNextBiomesAdapter :
+    ListAdapter<NextBiomeUiModel, BiomeNextBiomesViewHolder>(biomeComparator) {
     override fun onCreateViewHolder(
         parent: ViewGroup,
         viewType: Int,
-    ): BiomeMapViewHolder {
-        return BiomeMapViewHolder(
-            ItemBiomeMapBinding.inflate(
+    ): BiomeNextBiomesViewHolder {
+        return BiomeNextBiomesViewHolder(
+            ItemBiomeNextBiomesBinding.inflate(
                 LayoutInflater.from(parent.context),
                 parent,
                 false,
@@ -23,7 +23,7 @@ class BiomeMapAdapter :
     }
 
     override fun onBindViewHolder(
-        holder: BiomeMapViewHolder,
+        holder: BiomeNextBiomesViewHolder,
         position: Int,
     ) {
         holder.bind(getItem(position))
