+++ android/app/src/main/res/layout/fragment_biome_next_biome.xml
@@ -18,7 +18,7 @@
             android:layout_height="match_parent"
             app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
             app:layout_constraintTop_toTopOf="parent"
-            tools:listitem="@layout/item_biome_map" />
+            tools:listitem="@layout/item_biome_next_biomes" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
